requirements.txt
boto==2.49.0
boto3==1.9.236
botocore==1.12.236
dj-database-url==0.5.0
Django==2.1
django-crispy-forms==1.7.2
django-heroku==0.3.1
django-storages==1.7.2
docutils==0.15.2
gunicorn==19.9.0
jmespath==0.9.4
joblib==0.13.2

numpy==1.17.2
pandas==0.25.1
Pillow==6.1.0
psycopg2==2.8.3
python-dateutil==2.8.0
pytz==2019.2
s3transfer==0.2.1
scikit-learn==0.21.3
scipy==1.3.1
six==1.12.0
sklearn==0.0
urllib3==1.25.6
whitenoise==4.1.4
djangorestframework==3.11.0
